pyproject.toml
[project]
name = "pet-recommendation-system"
version = "0.1.0"
description = "A system for recommending pets based on user preferences." # Более описательное описание
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" } # Рекомендуется указывать лицензию
keywords = ["pets", "recommendation", "fastapi", "dynamodb"] # Добавьте ключевые слова

dependencies = [
    "aioboto3>=15.0.0",
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "uvicorn>=0.35.0",
    "python-dotenv>=1.0.0",
    "types-aioboto3>=15.0.0",
    "pre-commit>=4.2.0",
]

[project.optional-dependencies]
dev = [
    "bandit>=1.8.6", # Обновленная версия
    "black>=25.1.0", # Обновленная версия
    "flake8>=7.3.0", # Обновленная версия
    "isort>=6.0.1",  # Обновленная версия
    "mypy>=1.16.1",  # Обновленная версия
    "pre-commit>=4.2.0", # Обновленная версия
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \bbuild\b
  | \bdist\b
)/
'''


[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]


[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
